requirements.txt
PyQt6==6.7.0
Pillow==12.1.1
